fix: move some stuff around
