+++ src/services/thema.ts
@@ -14,25 +14,23 @@ import Ouderen from "../themas/Ouderen";
 import Veiligheid from "../themas/Veiligheid";
 import Zorg from "../themas/Zorg";
 
-/**
- * Each thema is identied by a string in the UI
- */
-
-export const IN_HET_KORT = "Gebied in het kort";
-export const BEVOLKING = "Bevolking";
-export const DUURZAAMHEID = "Duurzaamheid";
-export const ECONOMIE = "Economie";
-export const ONDERWIJS = "Onderwijs";
-export const VERKEER = "Verkeer";
-export const SPORT_EN_CULTUUR = "Sport en cultuur";
-export const WERK_INKOMEN = "Werk en inkomen";
-export const WONEN = "Wonen";
-export const SOCIALE_KRACHT = "Sociale kracht";
-export const OPENBARE_RUIMTE = "Openbare ruimte";
-export const VEILIGHEID = "Veiligheid en Overlast";
-export const GEZONDHEID_ZORG = "Gezondheid en zorg";
-export const JEUGD = "Jeugd";
-export const OUDEREN = "Ouderen";
+import {
+  BEVOLKING,
+  DUURZAAMHEID,
+  ECONOMIE,
+  GEZONDHEID_ZORG,
+  IN_HET_KORT,
+  JEUGD,
+  ONDERWIJS,
+  OPENBARE_RUIMTE,
+  OUDEREN,
+  SOCIALE_KRACHT,
+  SPORT_EN_CULTUUR,
+  VEILIGHEID,
+  VERKEER,
+  WERK_INKOMEN,
+  WONEN,
+} from "./themaNames";
 
 /**
  * The total set of available themas of which the user can choose from
