+++ src/services/themaNames.ts
@@ -1,0 +1,19 @@
+/**
+ * Each thema is identied by a string in the UI
+ */
+
+export const IN_HET_KORT = "Gebied in het kort";
+export const BEVOLKING = "Bevolking";
+export const DUURZAAMHEID = "Duurzaamheid";
+export const ECONOMIE = "Economie";
+export const ONDERWIJS = "Onderwijs";
+export const VERKEER = "Verkeer";
+export const SPORT_EN_CULTUUR = "Sport en cultuur";
+export const WERK_INKOMEN = "Werk en inkomen";
+export const WONEN = "Wonen";
+export const SOCIALE_KRACHT = "Sociale kracht";
+export const OPENBARE_RUIMTE = "Openbare ruimte";
+export const VEILIGHEID = "Veiligheid en Overlast";
+export const GEZONDHEID_ZORG = "Gezondheid en zorg";
+export const JEUGD = "Jeugd";
+export const OUDEREN = "Ouderen";
